pyproject.toml
[project]
name = "data-engineering-zoomcamp"
version = "0.1.0"
description = "Data Engineering Zoomcamp - Shared environment for all modules"
requires-python = ">=3.13"
dependencies = [
    # Module 01 - Docker & Terraform
    "click>=8.1.7",
    "pandas>=3.0.0",
    "psycopg2-binary>=2.9.11",
    "pyarrow>=23.0.0",
    "sqlalchemy>=2.0.46",
    "tqdm>=4.67.1",
    # Module 03 - Data Warehouse
    "google-cloud-storage",
    # Module 04 - Analytics Engineering
    "duckdb",
    "dbt-duckdb>=1.10.0",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "pgcli>=4.4.0",
]
